requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

boto3==1.34.53
python-dotenv==1.0.1
httpx==0.27.0
openai==1.14.2

pillow==10.2.0
pillow-heif==0.13.1
